refactor(debt): extract runtime/extraction.rs; runtime.rs under 1500 LOC (D-01 step 3)
Move extract_runtime + check_extraction_progress + is_ready_without_schema_check
+ is_runtime_ready + write_schema_version into runtime/extraction.rs. Shared
helpers (runtime_dir, native_lib_dir, check_health), the RUNTIME_* consts and
the ExtractionProgress payload stay in runtime.rs, reached via use super::*.

runtime.rs 1592 -> 1491 LOC, now UNDER the 1500 LOC gate. The god *file* is
resolved; only the god *function* start_embedded_server (669 LOC) remains, which
needs device-in-loop validation to relocate AND decompose together (step 2).

extract_runtime re-export is #[allow(unused_imports)] because lib.rs's
generate_handler! consuming it is #[cfg(target_os=android)] (unused on host).

Verified under WSL Linux: cargo test --lib runtime 14/0; clippy --lib --tests
-D warnings clean.
